Reverted r479 since SimonUtils.presentNanoTime behaviour was reverted in r481 Improved Stringifier to take advantage of SimonUtils.presentNanoTime(double)

+++ console/embed/src/main/java/org/javasimon/console/text/StringifierFactory.java
@@ -232,7 +232,7 @@ public class StringifierFactory {
 		@Override
 		protected String doToString(Double d) {
 			if (timeFormatType == TimeFormatType.AUTO) {
-				return stringStringifier.toString(SimonUtils.presentNanoTime(d.longValue()));
+				return stringStringifier.toString(SimonUtils.presentNanoTime(d));
 			} else {
 				Double d2 = timeFormatType.convert(d);
 				return doubleStringifier.toString(d2);
